Fixed compilation problem in Linux
Former-commit-id: a6ba934fb340b3e2bde30c79d66d3a5c6ec8b2eb

+++ modules/decklink/consumer/decklink_consumer.cpp
@@ -587,7 +587,7 @@ public:
 
 	int64_t presentation_frame_age_millis() const override
 	{
-		return consumer_ ? consumer_->current_presentation_delay_ : 0;
+		return consumer_ ? static_cast<int64_t>(consumer_->current_presentation_delay_) : 0;
 	}
 
 	core::monitor::subject& monitor_output()
